+++ core/src/commonMain/kotlin/org/yapyap/config/TransportLimits.kt
@@ -1,16 +1,16 @@
 package org.yapyap.config
 
 data class TransportLimits(
-    val torMaxPayloadBytes: Long,
-    val webRtcMaxPayloadBytes: Long,
+    val torMaxPayloadBytes: Int,
+    val webRtcMaxPayloadBytes: Int,
 ) {
     /** Max BinaryEnvelope that can traverse EITHER transport. Messages use this. */
-    val maxRoutableBytes: Long
+    val maxRoutableBytes: Int
         get() = minOf(torMaxPayloadBytes, webRtcMaxPayloadBytes)
 
     /** Max BinaryEnvelope that can traverse the LARGER transport.
      *  Unchunked file envelopes use this. Crypto DoS guard uses this. */
-    val maxTransportableBytes: Long
+    val maxTransportableBytes: Int
         get() = maxOf(torMaxPayloadBytes, webRtcMaxPayloadBytes)
 
     companion object {
